+++ docs/source/changelog.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## v0.x.x (Unreleased)
+### New features
+
+### Maintenance and fixes
+
+### Documentation
+
 ## v0.5.0 (2023 Jan 16)
 ### New features
 * Added {func}`.empty_ref`, {func}`.ones_ref` and {func}`.zeros_ref` DataArray creation helpers {pull}`37`